Remove int64_t from implementation

+++ src/mutable_path_metadata.cpp
@@ -11,8 +11,8 @@ namespace handlegraph {
 path_handle_t MutablePathMetadata::create_path(const PathSense& sense,
                                                const std::string& sample,
                                                const std::string& locus,
-                                               const int64_t& haplotype,
-                                               const int64_t& phase_block,
+                                               const size_t& haplotype,
+                                               const size_t& phase_block,
                                                const subrange_t& subrange,
                                                bool is_circular) {
